move social history service functions out of evaluateFhirDataService. Move and update imports and tests
