+++ containers/ecr-viewer/src/app/view-data/components/EcrDocument/accordion-items.tsx
@@ -15,13 +15,13 @@ import UnavailableInfo from "@/app/view-data/components/UnavailableInfo";
 import { evaluateEcrMetadata } from "@/app/view-data/services/ecrMetadataService";
 import {
   evaluateDemographicsData,
-  evaluateSocialData,
   evaluateEncounterData,
   evaluateProviderData,
   evaluateFacilityData,
   evaluateHospitalEncounterData,
   evaluatePregnancyData,
 } from "@/app/view-data/services/evaluateFhirDataService";
+import { evaluateSocialData } from "@/app/view-data/services/socialHistoryService";
 import { evaluateLabInfoData } from "@/app/view-data/services/labsService";
 import {
   FhirIndex,
