+++ containers/ecr-viewer/src/app/view-data/services/evaluateFhirDataService.tsx
@@ -24,7 +24,6 @@ import { ExpandCollapseAccordion } from "@/app/components/ExpandCollapseAccordio
 import {
   formatDate,
   formatDateTime,
-  formatPeriodDate,
   formatStartEndDate,
   formatStartEndDateTime,
 } from "@/app/services/formatDateService";
@@ -38,7 +37,6 @@ import {
   formatPatientContactList,
   formatAge,
   findCurrentAddress,
-  formatCurrentAddress,
   Age,
   formatCoding,
 } from "@/app/services/formatService";
@@ -60,7 +58,7 @@ import {
   evaluateValue,
 } from "@/app/utils/evaluate";
 import fhirPathMappings from "@/app/utils/evaluate/fhir-paths";
-import { toSentenceCase, toTitleCase } from "@/app/utils/format-utils";
+import { toTitleCase } from "@/app/utils/format-utils";
 import {
   DataDisplay,
   DataDisplayList,
@@ -76,12 +74,6 @@ import {
   sortResourcesByDate,
 } from "@/app/view-data/utils/fhir-data-utils";
 
-import {
-  evaluateExposureDetails,
-  evaluateTravelHistoryTable,
-  returnDisabilityStatusTable,
-  evaluateSocialDeterminantsOfHealth,
-} from "./socialHistoryService";
 import {
   FhirIndex,
   getOneResourceByType,
@@ -478,317 +470,6 @@ export const evaluateDemographicsData = (
     },
   ];
   return evaluateData(demographicsData);
-};
-
-// =============================================================================
-// Patient Info: Social History
-// =============================================================================
-
-/**
- * Evaluates alcohol use information from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing alcohol use data.
- * @returns An array of evaluated and formatted alcohol use data.
- */
-export const evaluateAlcoholUse = (fhirBundle: Bundle) => {
-  const alcoholUse = evaluateValue(
-    fhirBundle,
-    fhirPathMappings.patientAlcoholUse,
-  );
-  const alcoholIntake = evaluateValue(
-    fhirBundle,
-    fhirPathMappings.patientAlcoholIntake,
-  );
-  let alcoholComment: string | undefined = evaluateValue(
-    fhirBundle,
-    fhirPathMappings.patientAlcoholComment,
-  );
-
-  if (alcoholComment) {
-    alcoholComment = toSentenceCase(alcoholComment);
-  }
-
-  return [
-    alcoholUse ? `Use: ${alcoholUse}` : null,
-    alcoholIntake ? `Intake (standard drinks/week): ${alcoholIntake}` : null,
-    alcoholComment ? `Comment: ${alcoholComment}` : null,
-  ]
-    .filter(Boolean) // Removes null or undefined lines
-    .join("\n"); // Joins the remaining lines with newlines
-};
-
-/**
- * Evaluates occupation information from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing alcohol use data.
- * @returns An array of evaluated and formatted occupation data.
- */
-export const evaluateOccupation = (fhirBundle: Bundle) => {
-  const occupationObs = evaluateOne(
-    fhirBundle,
-    fhirPathMappings.patientOccupationUsual,
-  );
-  const employmentObs = evaluateAll(
-    fhirBundle,
-    fhirPathMappings.patientEmploymentStatus,
-  );
-  if (!occupationObs && employmentObs.length === 0) return;
-
-  const occTitle = formatCodeableConcept(occupationObs?.valueCodeableConcept);
-  const occDates = formatPeriodDate(occupationObs?.effectivePeriod);
-  const usualIndustryComp = occupationObs?.component?.find(
-    (c) => c?.code.coding?.[0].code === "21844-6",
-  );
-  const usualIndustry = formatCodeableConcept(
-    usualIndustryComp?.valueCodeableConcept,
-  );
-
-  sortResourcesByDate(employmentObs, fhirPathMappings.effectiveX);
-  const currentEmploymentStatus = evaluateValue(
-    employmentObs,
-    fhirPathMappings.valueX,
-  );
-
-  return [
-    occTitle,
-    usualIndustry && `Industry: ${usualIndustry}`,
-    currentEmploymentStatus && `Status: ${currentEmploymentStatus}`,
-    occDates && `Dates: ${occDates}`,
-  ]
-    .filter(Boolean)
-    .join("\n\n");
-};
-
-/**
- * Evaluates occupation history information from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing alcohol use data.
- * @returns An array of evaluated and formatted occupation history data.
- */
-export const evaluateOccupationHistory = (fhirBundle: Bundle) => {
-  const jobObs = [
-    // Employment detail(s) may come from a Past or Present Occupation Obs
-    ...evaluateAll(
-      fhirBundle,
-      fhirPathMappings.patientOccupationFromPastOrPresent,
-    ),
-    // or a Social History Obs
-    ...evaluateAll(
-      fhirBundle,
-      fhirPathMappings.patientOccupationFromSocialHistory,
-    ),
-  ];
-  if (jobObs.length === 0) return;
-
-  sortResourcesByDate(jobObs, fhirPathMappings.effectiveX);
-
-  return (
-    <ExpandCollapseAccordion
-      className="accordion-rr"
-      descriptor="employment details"
-      items={jobObs.map((obs) => {
-        const getComponentValue = (code: string) => {
-          return (
-            evaluateValue(
-              obs,
-              `component.where(code.coding.code = '${code}').value`,
-            ) || noData
-          );
-        };
-
-        const employerRef = evaluateValue(
-          obs,
-          "extension.where(url = 'http://hl7.org/fhir/us/odh/StructureDefinition/odh-Employer-extension').value",
-        );
-        const employer = evaluateReference<RelatedPerson | Organization>(
-          fhirBundle,
-          employerRef,
-        );
-
-        const workplaceInfo = [
-          {
-            title: "Address",
-            value: formatCurrentAddress(employer?.address) || noData,
-          },
-          {
-            title: "Schedule",
-            value: getComponentValue("74159-5"),
-          },
-          {
-            title: "Hours",
-            value: getComponentValue("87512-0"),
-          },
-          {
-            title: "Days",
-            value: getComponentValue("74160-3"),
-          },
-          {
-            title: "Duties",
-            value: getComponentValue("63761-1"),
-          },
-          {
-            title: "Pay Grade",
-            value: getComponentValue("87707-6"),
-          },
-          {
-            title: "Employment Type",
-            value: getComponentValue("85104-8"),
-          },
-        ];
-
-        const hasWorkplaceContent = workplaceInfo.some(
-          ({ value }) => value !== noData,
-        );
-
-        const workplaceContent = hasWorkplaceContent
-          ? workplaceInfo.map(({ title, value }, i) => (
-              <DataDisplay
-                key={`wi-${i}`}
-                item={{
-                  title,
-                  value,
-                  dividerLine: false,
-                  titleNormal: true,
-                }}
-              />
-            ))
-          : noData;
-
-        const content = (
-          <>
-            <DataDisplay
-              item={{
-                title: "Dates",
-                value: obs.effectivePeriod
-                  ? formatPeriodDate(obs.effectivePeriod)
-                  : formatDateTime(obs.effectiveDateTime),
-              }}
-            />
-            <DataDisplay
-              item={{
-                title: "Industry",
-                value: getComponentValue("86188-0"),
-              }}
-            />
-            <DataDisplay
-              item={{
-                title: "Workplace Information",
-                value: workplaceContent,
-                fullWidthContent: hasWorkplaceContent,
-              }}
-            />
-            <DataDisplay
-              item={{
-                title: "Hazard",
-                value: getComponentValue("87729-0"),
-                dividerLine: false,
-              }}
-            />
-          </>
-        );
-
-        return {
-          title: (
-            <div className="display-flex flex-row flex-no-wrap flex-justify">
-              <span>
-                {formatCodeableConcept(obs.valueCodeableConcept) ??
-                  obs.valueString}
-              </span>
-              <span className="font-size-xs text-base">
-                {!!obs.effectivePeriod?.end ? "Past" : "Current"} Employment
-              </span>
-            </div>
-          ),
-          expanded: false,
-          content,
-          id: obs.id || `${Math.random()}`,
-          headingLevel: "h5",
-        };
-      })}
-    />
-  );
-};
-
-/**
- * Evaluates social data from the FHIR bundle and formats it into structured data for display.
- * @param fhirBundle - The FHIR bundle containing social data.
- * @returns An array of evaluated and formatted social data.
- */
-export const evaluateSocialData = (
-  fhirBundle: Bundle,
-  fhirIndex: FhirIndex,
-) => {
-  const patient = getPatient(fhirIndex);
-
-  const socialData: DisplayDataProps[] = [
-    {
-      title: "Exposure Contacts",
-      value: evaluateExposureDetails(fhirBundle),
-      fullWidthContent: true,
-    },
-    {
-      title: "Tobacco Use",
-      value: evaluateValue(fhirBundle, fhirPathMappings.patientTobaccoUse),
-    },
-    {
-      title: "Travel History",
-      value: evaluateTravelHistoryTable(fhirBundle),
-      fullWidthContent: true,
-    },
-    {
-      title: "Homeless Status",
-      value: evaluateValue(fhirBundle, fhirPathMappings.patientHomelessStatus),
-    },
-    {
-      title: "Alcohol Use",
-      value: evaluateAlcoholUse(fhirBundle),
-    },
-    {
-      title: "Sexual Orientation",
-      value: evaluateValue(
-        fhirBundle,
-        fhirPathMappings.patientSexualOrientation,
-      ),
-    },
-    {
-      title: "Occupation",
-      value: evaluateOccupation(fhirBundle),
-    },
-    {
-      title: "Occupation History",
-      value: evaluateOccupationHistory(fhirBundle),
-      fullWidthContent: true,
-    },
-    {
-      title: "Religious Affiliation",
-      value: evaluateValue(patient, fhirPathMappings.patientReligion),
-    },
-    {
-      title: "Marital Status",
-      value: evaluateValue(patient, fhirPathMappings.patientMaritalStatus),
-    },
-    {
-      title: "Nationality",
-      value: evaluateValue(fhirBundle, fhirPathMappings.patientNationality),
-    },
-    {
-      title: "Country of Residence",
-      value: evaluateValue(
-        fhirBundle,
-        fhirPathMappings.patientCountryResidence,
-      ),
-    },
-    {
-      title: "Disability Status",
-      value: returnDisabilityStatusTable(fhirBundle),
-      fullWidthContent: true,
-    },
-    {
-      title: "Social Determinants of Health",
-      value: evaluateSocialDeterminantsOfHealth(fhirBundle),
-      fullWidthContent: true,
-      fullWidthTitle: true,
-    },
-  ];
-
-  return evaluateData(socialData);
 };
 
 // =============================================================================
